修复查看订单和订单详情中的bug

+++ src/main/java/Server/pojo/Order.java
@@ -19,7 +19,7 @@ public class Order {
     private String orderCode;
     private String receiveAddr;
     private String receiveName;
-    private int telPhone;
+    private long telPhone;
     private double price;
     private int status;
     private Timestamp gmtCreate;
